Extract config readers to separate module
Because:
- portadmin is weirdly split into the nav.portadmin and
  nav.web.portadmin packages.
- Config parsing code was kept in nav.web.portadmin, but will soon be
  needed also in nav.portadmin.
- It's, IMHO, best to coalesce everything config-parsing related into a
  module of its own.
- There may be import cycles if nav.portadmin modules begin importing
  stuff from the nav.web.portadmin package.
